refactor: move multiple items around into more proper places
