+++ mysk-data-api/src/routes/v1/subjects/electives/enroll_electives.rs
@@ -9,7 +9,7 @@ use actix_web::{
 };
 use mysk_lib::{
     common::{
-        requests::{FetchLevel, QueryablePlaceholder, RequestType, SortablePlaceholder},
+        requests::{FetchLevel, RequestType, SortablePlaceholder},
         response::ResponseType,
     },
     helpers::date::{get_current_academic_year, get_current_semester},
@@ -19,6 +19,7 @@ use mysk_lib::{
     },
     permissions,
     prelude::*,
+    query::QueryablePlaceholder,
 };
 use sqlx::query;
 use uuid::Uuid;
