+++ mysk-lib/src/common/mod.rs
@@ -1,4 +1,7 @@
 pub mod config;
+pub mod pagination;
 pub mod requests;
 pub mod response;
 pub mod string;
+
+pub use pagination::{PaginationConfig, PaginationType};
