+++ mysk-lib/src/common/requests.rs
@@ -1,25 +1,51 @@
-use crate::{
-    models::{
-        enums::{ContactType, ShirtSize, SubmissionStatus},
-        traits::Queryable,
-    },
-    prelude::*,
-};
+use crate::{common::PaginationConfig, prelude::*, query::Queryable};
 use actix_web::{dev::Payload, FromRequest, HttpRequest};
-use chrono::NaiveDate;
 use futures::future;
-use serde::{de::DeserializeOwned, Deserialize, Serialize};
-use sqlx::{Encode, Postgres, QueryBuilder};
+use serde::{de::DeserializeOwned, Deserialize};
+use sqlx::{Postgres, QueryBuilder};
 use std::fmt::{Display, Formatter};
 use std::string::ToString;
-use uuid::Uuid;
 
-#[derive(Debug, Deserialize)]
-pub struct QueryablePlaceholder;
+#[derive(Clone, Copy, Debug, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum FetchLevel {
+    IdOnly,
+    Compact,
+    Default,
+    Detailed,
+}
 
-impl Queryable for QueryablePlaceholder {
-    fn to_query_string(&self) -> Vec<SqlSection> {
-        unimplemented!("QueryablePlaceholder can't actually be queried");
+#[derive(Clone, Debug, Deserialize)]
+pub struct FilterConfig<T> {
+    pub data: Option<T>,
+    pub q: Option<String>,
+}
+
+#[derive(Clone, Debug, Deserialize)]
+pub struct SortingConfig<S: Display> {
+    by: Vec<S>,
+    ascending: Option<bool>,
+}
+
+impl<S: Display> SortingConfig<S> {
+    pub fn new(by: Vec<S>, ascending: Option<bool>) -> Self {
+        Self { by, ascending }
+    }
+
+    pub fn append_into_query_builder(&self, qb: &mut QueryBuilder<'_, Postgres>) {
+        qb.push(" ORDER BY ")
+            .push(
+                self.by
+                    .iter()
+                    .map(ToString::to_string)
+                    .collect::<Vec<String>>()
+                    .join(", "),
+            )
+            .push(if self.ascending.unwrap_or(true) {
+                " ASC"
+            } else {
+                " DESC"
+            });
     }
 }
 
@@ -32,111 +58,22 @@ impl Display for SortablePlaceholder {
     }
 }
 
-#[derive(Clone, Copy, Debug, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum FetchLevel {
-    IdOnly,
-    Compact,
-    Default,
-    Detailed,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct FilterConfig<T> {
-    pub data: Option<T>,
-    pub q: Option<String>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct SortingConfig<T> {
-    pub by: Vec<T>,
-    pub ascending: Option<bool>,
-}
-
-impl<SortingObject> SortingConfig<SortingObject>
-where
-    SortingObject: Display,
-{
-    pub fn new(by: Vec<SortingObject>, ascending: Option<bool>) -> Self {
-        Self { by, ascending }
-    }
-
-    pub fn to_order_by_clause(&self) -> String {
-        let mut order_by = " ORDER BY ".to_string();
-        let columns = self
-            .by
-            .iter()
-            .map(ToString::to_string)
-            .collect::<Vec<String>>()
-            .join(", ");
-        order_by.push_str(&columns);
-
-        if let Some(ascending) = self.ascending {
-            order_by.push_str(if ascending { " ASC" } else { " DESC" });
-        }
-
-        order_by
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct PaginationConfig {
-    pub p: u32,
-    pub size: Option<u32>,
-}
-
-impl Default for PaginationConfig {
-    fn default() -> Self {
-        Self {
-            p: 1,
-            size: Some(50),
-        }
-    }
-}
-
-impl PaginationConfig {
-    pub fn new(p: u32, size: Option<u32>) -> Self {
-        Self { p, size }
-    }
-
-    pub fn to_limit_clause(&self) -> Result<SqlSection> {
-        if self.p == 0 {
-            return Err(Error::InvalidRequest(
-                "Page number must be greater than zero".to_string(),
-                "PaginationConfig::to_limit_clause".to_string(),
-            ));
-        }
-
-        // LIMIT $1 OFFSET $2
-        Ok(SqlSection {
-            sql: vec!["LIMIT ".to_string(), " OFFSET ".to_string()],
-            params: vec![
-                QueryParam::Int(i64::from(self.size.unwrap_or(50))),
-                QueryParam::Int(i64::from((self.p - 1) * self.size.unwrap_or(50))),
-            ],
-        })
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct RequestType<T, Queryable, Sortable>
-where
-    Sortable: Display,
-{
+#[derive(Clone, Debug, Deserialize)]
+pub struct RequestType<T, Q: Queryable, S: Display> {
     pub data: Option<T>,
     pub pagination: Option<PaginationConfig>,
-    pub filter: Option<FilterConfig<Queryable>>,
-    pub sort: Option<SortingConfig<Sortable>>,
+    pub filter: Option<FilterConfig<Q>>,
+    pub sort: Option<SortingConfig<S>>,
     pub fetch_level: Option<FetchLevel>,
     pub descendant_fetch_level: Option<FetchLevel>,
 }
 
-// Implement from request for RequestType with any T, Queryable, and Sortable
-impl<T, Queryable, Sortable> FromRequest for RequestType<T, Queryable, Sortable>
+// Implement from request for `RequestType` with any `T`, `Q`, and `S`
+impl<T, Q, S> FromRequest for RequestType<T, Q, S>
 where
     T: DeserializeOwned,
-    Queryable: DeserializeOwned,
-    Sortable: DeserializeOwned + Display,
+    Q: DeserializeOwned + Queryable,
+    S: DeserializeOwned + Display,
 {
     type Error = Error;
     type Future = future::Ready<Result<Self>>;
@@ -144,97 +81,11 @@ where
     fn from_request(req: &HttpRequest, _: &mut Payload) -> Self::Future {
         let query_string = req.query_string();
         let qs_parser = serde_qs::Config::new(5, false);
-        let request_query =
-            qs_parser.deserialize_str::<RequestType<T, Queryable, Sortable>>(query_string);
+        let request_query = qs_parser.deserialize_str::<RequestType<T, Q, S>>(query_string);
 
         match request_query {
             Ok(query) => future::ok(query),
             Err(e) => future::err(Error::InvalidRequest(e.to_string(), req.path().to_string())),
         }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-pub enum QueryParam {
-    Int(i64),
-    Float(f64),
-    String(String),
-    Bool(bool),
-    Uuid(Uuid),
-    NaiveDate(NaiveDate),
-    ArrayInt(Vec<i64>),
-    ArrayFloat(Vec<f64>),
-    ArrayString(Vec<String>),
-    ArrayBool(Vec<bool>),
-    ArrayUuid(Vec<Uuid>),
-    ArrayNaiveDate(Vec<NaiveDate>),
-    ContactType(ContactType),
-    ShirtSize(ShirtSize),
-    SubmissionStatus(SubmissionStatus),
-}
-
-impl QueryParam {
-    pub fn push_bind(self, qb: &mut QueryBuilder<'_, Postgres>) {
-        match self {
-            QueryParam::Int(v) => qb.push_bind(v),
-            QueryParam::Float(v) => qb.push_bind(v),
-            QueryParam::String(v) => qb.push_bind(v),
-            QueryParam::Bool(v) => qb.push_bind(v),
-            QueryParam::Uuid(v) => qb.push_bind(v),
-            QueryParam::NaiveDate(v) => qb.push_bind(v),
-            QueryParam::ArrayInt(v) => qb.push_bind(v),
-            QueryParam::ArrayFloat(v) => qb.push_bind(v),
-            QueryParam::ArrayString(v) => qb.push_bind(v),
-            QueryParam::ArrayBool(v) => qb.push_bind(v),
-            QueryParam::ArrayUuid(v) => qb.push_bind(v),
-            QueryParam::ContactType(v) => qb.push_bind(v),
-            QueryParam::ShirtSize(v) => qb.push_bind(v),
-            QueryParam::SubmissionStatus(v) => qb.push_bind(v),
-        };
-    }
-}
-
-impl Encode<'_, Postgres> for QueryParam {
-    fn encode_by_ref(
-        &self,
-        buf: &mut <Postgres as sqlx::Database>::ArgumentBuffer<'_>,
-    ) -> Result<sqlx::encode::IsNull, Box<dyn std::error::Error + Send + Sync + 'static>> {
-        match self {
-            QueryParam::String(v) => <String as sqlx::Encode<Postgres>>::encode(v.to_string(), buf),
-            QueryParam::Int(v) => <i64 as sqlx::Encode<Postgres>>::encode(*v, buf),
-            QueryParam::Float(v) => <f64 as sqlx::Encode<Postgres>>::encode(*v, buf),
-            QueryParam::Bool(v) => <bool as sqlx::Encode<Postgres>>::encode(*v, buf),
-            QueryParam::Uuid(v) => <Uuid as sqlx::Encode<Postgres>>::encode(*v, buf),
-            QueryParam::NaiveDate(v) => <NaiveDate as sqlx::Encode<Postgres>>::encode(*v, buf),
-            QueryParam::ArrayInt(v) => v.encode_by_ref(buf),
-            QueryParam::ArrayFloat(v) => v.encode_by_ref(buf),
-            QueryParam::ArrayString(v) => v.encode_by_ref(buf),
-            QueryParam::ArrayBool(v) => v.encode_by_ref(buf),
-            QueryParam::ArrayUuid(v) => v.encode_by_ref(buf),
-            QueryParam::ArrayNaiveDate(v) => v.encode_by_ref(buf),
-            QueryParam::ContactType(v) => <ContactType as sqlx::Encode<Postgres>>::encode(*v, buf),
-            QueryParam::ShirtSize(v) => <ShirtSize as sqlx::Encode<Postgres>>::encode(*v, buf),
-            QueryParam::SubmissionStatus(v) => {
-                <SubmissionStatus as sqlx::Encode<Postgres>>::encode(*v, buf)
-            }
-        }
-    }
-}
-
-// #[deprecated(since = "0.6.0")]
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct SqlSection {
-    pub sql: Vec<String>,
-    pub params: Vec<QueryParam>,
-}
-
-impl SqlSection {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn add(&mut self, sql: &str, param: QueryParam) {
-        self.sql.push(sql.to_string());
-        self.params.push(param);
     }
 }
