+++ mysk-lib/src/common/response.rs
@@ -1,66 +1,13 @@
+use crate::common::PaginationType;
 use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use sqlx::{query, PgPool};
-use std::fmt::{Display, Formatter};
 use uuid::Uuid;
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct PaginationType {
-    first_p: u32,
-    last_p: u32,
-    next_p: Option<u32>,
-    prev_p: Option<u32>,
-    size: u32,
-    total: u32,
-}
-
-impl PaginationType {
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    pub fn new(current_p: u32, size: u32, total: u32) -> Self {
-        let page_count = (f64::from(total) / f64::from(size)).ceil() as u32;
-
-        PaginationType {
-            first_p: 1,
-            last_p: page_count,
-            next_p: if current_p < page_count {
-                Some(current_p + 1)
-            } else {
-                None
-            },
-            prev_p: if current_p > 1 {
-                Some(current_p - 1)
-            } else {
-                None
-            },
-            size,
-            total,
-        }
-    }
-}
-
-impl Display for PaginationType {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "{{ first: {}, last: {}, next: {:?}, prev: {:?}, size: {}, total: {} }}",
-            self.first_p, self.last_p, self.next_p, self.prev_p, self.size, self.total,
-        )
-    }
-}
 
 #[derive(Debug, Serialize)]
 pub struct MetadataType {
     timestamp: DateTime<Utc>,
     pagination: Option<PaginationType>,
-}
-
-impl MetadataType {
-    pub fn new(pagination: Option<PaginationType>) -> Self {
-        MetadataType {
-            timestamp: Utc::now(),
-            pagination,
-        }
-    }
 }
 
 impl Default for MetadataType {
@@ -72,13 +19,12 @@ impl Default for MetadataType {
     }
 }
 
-impl Display for MetadataType {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "{{ timestamp: {}, pagination: {:?} }}",
-            self.timestamp, self.pagination,
-        )
+impl MetadataType {
+    pub fn new(pagination: Option<PaginationType>) -> Self {
+        MetadataType {
+            timestamp: Utc::now(),
+            pagination,
+        }
     }
 }
 
@@ -116,7 +62,7 @@ pub struct ErrorType {
 }
 
 impl ErrorType {
-    // TODO
+    // TODO: tracing and error logging
     pub async fn log(&self, pool: &PgPool, api_key: Option<Uuid>) {
         query!(
             "
@@ -133,16 +79,6 @@ impl ErrorType {
         .execute(pool)
         .await
         .unwrap();
-    }
-}
-
-impl Display for ErrorType {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "{{ id: {}, code: {}, error_type: {}, detail: {}, source: {} }}",
-            self.id, self.code, self.error_type, self.detail, self.source,
-        )
     }
 }
 
@@ -164,15 +100,5 @@ impl ErrorResponseType {
             data: None,
             meta,
         }
-    }
-}
-
-impl Display for ErrorResponseType {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "{{ api_version: {}, error: {}, data: {:?}, meta: {:?} }}",
-            self.api_version, self.error, self.data, self.meta,
-        )
     }
 }
