+++ mysk-lib/src/query/mod.rs
@@ -1,7 +1,7 @@
-pub mod fragment;
+pub mod building_blocks;
 pub mod set_clause;
 pub mod where_clause;
 
-pub use fragment::QueryFragment;
+pub use building_blocks::{QueryFragment, QueryParam, Queryable, QueryablePlaceholder};
 pub use set_clause::SqlSetClause;
 pub use where_clause::SqlWhereClause;
